Chores: code split
