+++ src/components/Card.tsx
@@ -1,0 +1,13 @@
+import { Box, BoxProps, useColorModeValue } from "@chakra-ui/react";
+import React from "react";
+
+export const Card = (props: BoxProps) => (
+  <Box
+    bg={useColorModeValue("white", "gray.700")}
+    py="8"
+    px={{ base: "4", md: "10" }}
+    shadow="base"
+    rounded={{ sm: "lg" }}
+    {...props}
+  />
+);
