+++ src/components/SignIn.tsx
@@ -14,6 +14,7 @@ import React from "react";
 import { FaGoogle } from "react-icons/fa";
 
 import { signInWithGoogle } from "../service/firebase";
+import { Card } from "./Card";
 
 const _signInWithGoogle = () => signInWithGoogle();
 
@@ -61,17 +62,6 @@ export const Link = (props: HTMLChakraProps<"a">) => (
     color={useColorModeValue("blue.500", "blue.200")}
     _hover={{ color: useColorModeValue("blue.600", "blue.300") }}
     display={{ base: "block", sm: "inline" }}
-    {...props}
-  />
-);
-
-export const Card = (props: BoxProps) => (
-  <Box
-    bg={useColorModeValue("white", "gray.700")}
-    py="8"
-    px={{ base: "4", md: "10" }}
-    shadow="base"
-    rounded={{ sm: "lg" }}
     {...props}
   />
 );
